requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
watchdog>=3.0.0
requests>=2.31.0
urllib3>=2.0.0
